ISPN-9864 Move non-API classes from org.infinispan.query.remote.client to org.infinispan.query.remote.client.impl package
