+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/event/ContinuousQueryImpl.java
@@ -19,7 +19,7 @@ import org.infinispan.protostream.SerializationContext;
 import org.infinispan.query.api.continuous.ContinuousQuery;
 import org.infinispan.query.api.continuous.ContinuousQueryListener;
 import org.infinispan.query.dsl.Query;
-import org.infinispan.query.remote.client.ContinuousQueryResult;
+import org.infinispan.query.remote.client.impl.ContinuousQueryResult;
 
 /**
  * A container of continuous query listeners for a cache.
