+++ remote-query/remote-query-client/src/main/java/org/infinispan/query/remote/client/impl/BaseQueryResponse.java
@@ -1,4 +1,4 @@
-package org.infinispan.query.remote.client;
+package org.infinispan.query.remote.client.impl;
 
 import java.io.IOException;
 import java.util.List;
@@ -13,5 +13,4 @@ public interface BaseQueryResponse {
    List<?> extractResults(SerializationContext serializationContext) throws IOException;
 
    long getTotalResults();
-
 }
