+++ remote-query/remote-query-client/src/main/java/org/infinispan/query/remote/client/impl/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Hot Rod query client side implementation details. This package is strictly for internal use.
+ *
+ * @private
+ */
+package org.infinispan.query.remote.client.impl;
