+++ remote-query/remote-query-server/src/main/java/org/infinispan/query/remote/impl/BaseRemoteQueryManager.java
@@ -10,7 +10,7 @@ import org.infinispan.encoding.DataConversion;
 import org.infinispan.factories.ComponentRegistry;
 import org.infinispan.query.dsl.IndexedQueryMode;
 import org.infinispan.query.dsl.Query;
-import org.infinispan.query.remote.client.QueryRequest;
+import org.infinispan.query.remote.client.impl.QueryRequest;
 
 /**
  * @since 9.4
